vocab_extractor/pyproject.toml
[tool.poetry]
name = "vocab-extractor"
version = "0.1.0"
description = ""
authors = ["David J Nevin <dnevincodes@gmail.com>"]
readme = "README.md"
packages = [{include = "vocab_extractor"}]

[tool.poetry.dependencies]
python = "^3.11"
textract = "^1.6.5"
python-docx = "^0.8.11"
nltk = "^3.8.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
python-docx = "^0.8.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
